Rename socket to transport, delete client/vendor.js (unused)
