--- client/vendor.js
@@ -1,8 +1,0 @@
-require('jquery');
-require('redux');
-
-// JS
-window.$ = window.jQuery = $;
-
-const FastClick = require('fastclick');
-FastClick.attach(document.body);
